StandaloneSmmPkg: Read BFV address from HOB if present
The address of the Boot Firmware Volume is populated by the Standalone
MM IPL in the SMM_CORE_PRIVATE_DATA structure by default. This patch
adds support for retrieving this information from a
EFI_HOB_FIRMWARE_VOLUME in the Hoblist. This alternative approach is
used when the MM Foundation is initialised during the SEC phase on ARM
standard platforms.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>