Readme.md: Add Build Flags section
Adds a "Build Flags" section to the build instructions to explain any
build flags that may be passed to influence the build.

The final package will not have any build flags. A reasonable number are
temporarily used to test different flows until final decisions are made
as to which flow to keep.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>