PrmPkg/PrmConfigDxe: Improve runtime MMIO range conversion algorithm
Added an assert to verify that the base address and length are aligned
on the system defined page boundary. This is used to report invalid
ranges for OS accessibility.

Check the GCD memory type of an existing memory descriptor found for
the MMIO range physical base and length and modify the descriptor
as needed for the range to be accessible at OS runtime.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>